refactor: 세부 파서 함수들 /parsers 디렉터리로 분리
